refactor(cli): split git module into focused subcommands
Restructures the large git.rs file into separate modules for better maintainability
and code organization. Each git subcommand now has its own dedicated module with
focused responsibilities and cleaner separation of concerns.

Key structural changes:
- Extract AmendCommand to git/amend.rs with amendment file processing
- Extract CheckCommand to git/check.rs with validation logic and batching
- Extract CreatePrCommand to git/create_pr.rs with AI-powered PR generation
- Extract InfoCommand to git/info.rs with repository analysis capabilities
- Extract TwiddleCommand to git/twiddle.rs with contextual commit improvement
- Extract ViewCommand to git/view.rs with YAML repository output
- Promote parse_beta_header to public(crate) for cross-module usage

Module organization benefits:
- Improved code maintainability through logical separation
- Easier testing and debugging of individual commands
- Reduced cognitive load when working on specific functionality
- Better encapsulation of command-specific logic
- Simplified main git.rs with clear module structure

Each module maintains full functionality while providing cleaner interfaces
and better code organization for future development and maintenance.
